Create Links for both full data sets  to download
The data downloaded and used for the models is too large to upload. All of the data used can be accessed through these links to kaggle. 

They will contain the main csv files, images, training, testing, and validation sets.

In order to run the code file you will need to download them locally or use an api to call the data in.